% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PJFM_wrapper.R
\docType{data}
\name{control_list}
\alias{control_list}
\title{control_list}
\description{
This list contains a list of parameters specifying the joint frailty model.
}
\details{
\itemize{
  \item ID_name: the variable name indicating the patient ID in both
  recurrent events data and survival data.
  \item item_name: the variable name indicating the types of recurrent events
  in the recurrent events data.
  \item time_name: the variable name indicating the occurrence time in the
  recurrent events data.
  \item fix_cov: a set of variables names indicating the covariates of
  fixed-effects in the recurrent events submodel.
  If NULL, not baseline covariates are included.
  \item random_cov: a set of variables names indicating the covariates of
  random-effects in the recurrent events submodel.
  If NULL, not baseline covariates are included.
  \item recur_fix_time_fun: a function specifying the time-related basis functions (fixed-effects) in
  the recurrent events submodel.
  \item recur_ran_time_fun: a function specifying the time-related basis functions (random-effects) in
  the recurrent events submodel. If this is an intercept only function,
  then only a random intercept is included (i.e. a joint frailty model).
  \item surv_fix_time_fun: a log-hazard function for the survival submodel.
  \item surv_time_name the variable name for the survival time
  in the survival data.
  \item surv_status_name the variable name for the censoring indicator
  in the survival data.
  \item surv_cov a set of variables names specifying the baseline covariates
  in the survival submodel.
  \item n_points an integer indicating the numebr of nodes being used in
  the Gaussian quadrature.
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
