% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PJFM_wrapper.R
\name{PJFM_summary}
\alias{PJFM_summary}
\title{The function to get summary table of PJFM fit.}
\usage{
PJFM_summary(res = NULL)
}
\arguments{
\item{res}{a model fit returned by PJFM_fit; SE estimates are only
available for JFM, but not PJFM.}
}
\value{
return a data frame, which contains parameter estimates in both submodels.
}
\description{
The function is used to get summary table of PJFM fit.
}
\references{
Jiehuan Sun. "Dynamic Prediction with Penalized Joint Frailty Model of High-Dimensional Recurrent Event Data and a Survival Outcome".
}
