% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PJFM_wrapper.R
\docType{data}
\name{RecurData}
\alias{RecurData}
\title{Simulated Recurrent Events Data}
\format{
A data frame with 57582 rows and 3 variables
}
\usage{
data(PJFMdata)
}
\description{
This dataset contains recurrent events data.
}
\details{
\itemize{
  \item ID: patient ID
  \item feature_id: types of recurrent events
  \item time: occurrence time
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
