% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PJFM_wrapper.R
\docType{data}
\name{SurvData}
\alias{SurvData}
\title{Simulated Survival Data}
\format{
A data frame with 300 rows and 4 variables
}
\usage{
data(PJFMdata)
}
\description{
This dataset contains survival outcome.
}
\details{
\itemize{
  \item ID: patient ID
  \item fstat: censoring indicator
  \item ftime: survival time
  \item x: baseline covariates
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
